+++ packages/sanddance-app/src/sanddanceApp.tsx
@@ -14,6 +14,7 @@ import {
 } from '@msrvida/sanddance-explorer';
 import { DataSource, DataSourceSnapshot, InsightMap } from './types';
 import { DataSourcePicker } from './dataSourcePicker';
+import { downloadData } from './download';
 import { strings } from './language';
 
 import VegaDeckGl = SandDance.VegaDeckGl;
@@ -136,36 +137,6 @@ export class SandDanceApp extends React.Component<Props, State> {
         );
     }
 
-    //To do: download csv, json, or tsv
-    private downloadData(data: any, datatype: string) {
-        var re = /.(csv|tsv|json|topojson)/;
-        var filename = this.state.dataSource.displayName.replace(re, '') + '.' + datatype;
-
-        // Adapted from https://ourcodeworld.com/articles/read/189/how-to-create-a-file-and-generate-a-download-with-javascript-in-the-browser-without-a-server
-        var element = document.createElement('a');
-        //element.setAttribute('href', 'data:text/json;charset=utf-8,' + encodeURIComponent(String(data)));
-        element.setAttribute('download', filename);
-        element.setAttribute('style', 'position:fixed;top:0;left: 0;z-index:9;background:pink');
-        element.innerText = 'download';
-        //element.style.display = 'none';
-        document.body.appendChild(element);
-
-        dataURIToBlob(data, blob => {
-            element.href = URL.createObjectURL(blob);
-            console.log(element.href);
-            // you must revoke the object URL, 
-            //   but since we can't know when the download occured, we have to attach it on the click handler..
-            element.onclick = function () {
-                // ..and to wait a frame
-                requestAnimationFrame(function () {
-                    URL.revokeObjectURL(element.href);
-                });
-                document.body.removeChild(element);
-            };
-            element.click();
-        });
-    }
-
     updateExplorerViewerOptions(viewerOptions: Partial<types.ViewerOptions>) {
         this.viewerOptions = viewerOptions;
         this.explorer && this.explorer.updateViewerOptions(this.viewerOptions);
@@ -213,7 +184,7 @@ export class SandDanceApp extends React.Component<Props, State> {
                         this.load(this.state.dataSource, snapshotOnLoad && snapshotOnLoad.insight);
                         this.props.mounted(this);
                     }}
-                    datasetExportHandler={(data, datatype) => this.downloadData(data, datatype)}
+                    datasetExportHandler={(data, datatype) => downloadData(data, datatype, this.state.dataSource.displayName)}
                     datasetElement={(
                         <DataSourcePicker
                             dataSource={this.state.dataSource}
@@ -244,22 +215,5 @@ export class SandDanceApp extends React.Component<Props, State> {
                 </Explorer>
             </section >
         );
-    }
-}
-
-//from https://stackoverflow.com/a/37151835/620501
-function dataURIToBlob(binStr: string, callback: (blob: Blob) => void) {
-    try {
-        var len = binStr.length,
-            arr = new Uint8Array(len);
-
-        for (var i = 0; i < len; i++) {
-            arr[i] = binStr.charCodeAt(i);
-        }
-
-        callback(new Blob([arr]));
-    }
-    catch (e) {
-        console.log(e);
     }
 }
